Replace `chronicles_disable_thread_ids` with `chronicles_thread_ids` (#164)
* Replace `chronicles_disable_thread_ids` with `chronicles_thread_ids`

This more closely matches the other "boolean" options such as line
numbers and makes the configuration more uniform and allows enabling
thread id for each log line separately.

Compatibility is maintained with the old options.

At the same time, make a pass over the unit test suite to remove thread
ids and/or disable the tests that don't match - because of a
[bug](https://github.com/status-im/nim-testutils/pull/62) in
nim-testrunnner, tests were not being run at all for quite some time
leading to regressions outside of the scope of this change.

In order to enable the disabled tests, testrunner needs to be updated to
support "ignoring" parts of an output or using a custom verifier, which
caters to thread id and dates that change with every run - the currently
hardcoded version in testrunner no longer matches what the library does.

This change removes the call to `setFirstProperty` since there is no
natural place for such a call - custom streams now have to keep track of
this information themselves - this is a backwards-incompatible change.

Parts of the implementation is sourced from
https://github.com/status-im/nim-chronicles/pull/88

* deterministic order of scope properties

* one more

* fix thread option